feat: этап 2

+++ stats/stats-dto/src/main/java/ru/practicum/hit/HitShort.java
@@ -5,5 +5,5 @@ public interface HitShort {
 
     String getUri();
 
-    Integer getHits();
+    Long getHits();
 }
